requirements.txt
great_expectations
snowflake-connector-python
snowflake-sqlalchemy
python-dotenv